+++ frontend/miniapp/src/domain/pricing/PricingRepository.js
@@ -1,0 +1,15 @@
+import { NotImplementedError } from '../../shared/errors/index.js';
+
+export class PricingRepository {
+  async getPriceList() {
+    throw new NotImplementedError();
+  }
+
+  async getPricingRules() {
+    throw new NotImplementedError();
+  }
+
+  async getPriceModelById(priceModelId) {
+    throw new NotImplementedError();
+  }
+}